refactor: implement DeckListPage and WordListPage components and refactor difficulty pages
