+++ frontend/src/components/common/WordListPage.module.scss
